There were some methods in ObservableSet that returned an ObservableSet, but whose values were not unique, as required by the Set contract.  The overrides I deleted and the rest I moved up to ObservableCollection.
